Refactor ship and upgrade data into classes
This is a pretty large change, so bear with me. The goal is to have a
generic 'store' node which can be tailored to ships, upgrades, or
anything else that wants an icon grid. This is the first step to it: it
keeps the exact same functionality as before, but refactors in the
following ways:

* Splits "Store" out of "ShipStore"
* Makes an "item" class to represent something sold in a store
* Replaces the dicts used for rows of Ships and Upgrades with classes
  * These classes inherit from Item
  * Extreme cleverness is applied to avoid enumerating ship vars

This required a fair bit of debugging because the failure modes of this
type of reflection heavy code tend to be "engine breaks" rather than a
helpful error message.

One very happy side effect of this is you can now typehint in more
places, and use `.var` rather than `["var"]` which I'm starting to come
around to.